+++ src/app.components/common/Search.tsx
@@ -38,6 +38,9 @@ export const SearchWrapper = styled.div`
     width: 100%;
     height: 30px;
     background-color: rgba(149, 149, 149, 0.4);
+    color: rgba(255, 255, 255, 0.8);
+    font-size: 15px;
+    padding: 7px;
   }
   .search-button {
     background-color: rgba(124, 124, 124, 0.4);
