+++ src/app.feature/conjunctions/module/useInstance.ts
@@ -1,0 +1,15 @@
+export const useInstance = (instance) => {
+  const { allColumns } = instance
+
+  let rowSpanHeaders = []
+
+  allColumns.forEach((column, i) => {
+    const { id, enableRowSpan } = column
+
+    if (enableRowSpan !== undefined) {
+      rowSpanHeaders = [...rowSpanHeaders, { id, topCellValue: null, topCellIndex: 0 }]
+    }
+  })
+
+  Object.assign(instance, { rowSpanHeaders })
+}
