+++ src/app.feature/favorite/component/ConjuctionsFavorite.tsx
@@ -4,15 +4,15 @@ import styled from 'styled-components'
 import ConjuctionsFavoriteTable from './ConjuctionsFavroiteTable'
 
 const ConjuctionsFavorite = () => {
+  const [inputValue, setInputValue] = useState('')
   const {
     handleSubmit,
     register,
     formState: { errors },
   } = useForm()
-  const [inputValue, setInpuValue] = useState('')
 
   const handleFavorite = (filed: { favorite: string }) => {
-    setInpuValue(filed.favorite.trim())
+    setInputValue(filed.favorite.trim())
   }
 
   return (
@@ -24,7 +24,9 @@ const ConjuctionsFavorite = () => {
           {...register('favorite')}
           placeholder="SatName or NorId"
         />
-        <button type="submit">Search</button>
+        <button className="favorite-form-btn" type="submit">
+          Search
+        </button>
         {errors.favorite && <span>Favorite Field is Required</span>}
       </form>
       <ConjuctionsFavoriteTable inputValue={inputValue} />
@@ -36,10 +38,31 @@ export default ConjuctionsFavorite
 
 const StyledWrapper = styled.div`
   margin-top: 10px;
+  text-align: center;
   .favorite-form {
     margin-bottom: 10px;
+    width: 300px;
+    display: flex;
+    gap: 1rem;
     .favorite-form-input {
-      margin-right: 10px;
+      margin-right: 5px;
+      border: none;
+      border-radius: 5px;
+      width: 100%;
+      height: 30px;
+      background-color: rgba(149, 149, 149, 0.4);
+      color: rgba(255, 255, 255, 0.8);
+      font-size: 15px;
+      padding: 7px;
+      ::placeholder {
+        color: white;
+        opacity: 0.5;
+      }
+    }
+    .favorite-form-btn {
+      background-color: rgba(124, 124, 124, 0.4);
+      color: #e2e2e2;
+      border-radius: 5px;
     }
   }
 `
