+++ src/app.feature/favorite/module/arrToMap.ts
